requirements.txt
requests
supabase
python-dotenv
pytz